+++ cli/cli/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+## [Unreleased]
 ### Added
 - New Code Analyzer to return compile time error for async void Callable methods.
 - New Code Fixer to fix async void Callable methods on IDE.
@@ -14,16 +14,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New Code Fixer to Solve Microservice classes missing Attribute or Partial keyword. 
 - New Code Analyzer to Check if Microservice Callable Methods return are inside Microservice Scope (Needs to be enabled by adding `<BeamValidateCallableTypesExistInSharedLibraries>true</BeamValidateCallableTypesExistInSharedLibraries>` to MS C# project)
 - New Code Analyzer and Fixer for Microservice ID non matches the `<BeamId>` csproj property.
+- New Code Analyzer and Fixer for non-readonly static fields on Microservice classes.
+- New Microservice Client Code Generator for Unity that used OAPI for the generation.
 - Added support for generating `FDateTime` instead of `FString` in Unreal code generation.
-- Added `beam config --set [--no-overrides]` command to enable local overrides to config variables like `PID`. 
+- Added `beam config --set [--no-overrides]` command to enable local overrides to config variables like `PID`.
   The intended usage of this command is to allow a user to select their current realm WITHOUT changing the `configuration-defaults.json` file which is committed to version control.
-- Added `beam org realms` command that prints out a list of all available realms for the requesting user. 
-
+- Added `beam org realms` command that prints out a list of all available realms for the requesting user.
+ 
 ### Changed
 - Revise the categorization of all Blueprint nodes to enhance discoverability.
+- `OptionalString` overrides `.ToString()` for easier print debugging.
+- Changes on MicroserviceBootstrapper to allow the creation of OAPI document after building the Microservice
 - `beam me` command now also gives you back your active token information
 - `beam init -q --cid my-game --username my@email.com --password my_password` now honors the quiet flag correctly. It'll auto-select the realm as the oldest development realm.
-- `IAccessToken`, the interface representing a Beamable access/refresh token pair, now exposes the `IssuedAt`/`ExpiresIn` data in addition to the `ExpiresAt` date.  
+- `IAccessToken`, the interface representing a Beamable access/refresh token pair, now exposes the `IssuedAt`/`ExpiresIn` data in addition to the `ExpiresAt` date.
+
+### Fixed
+- Fixed an issue in which running `beam deploy release` when CID was an alias resulted in an error in execution.
+
+## [4.3.0] - 2025-05-08
+### Added
+- `.beamignore` files may be used to ignore services and storages from the `beam deploy` commands. 
+- Hidden `--ignore-beam-ids` option can be used to ignore beam ids similar in addition to `.beamignore` files. [#4019](https://github.com/beamable/BeamableProduct/issues/4019)
+
+### Fixed
+- `beam deploy` commands no longer attempt to build non-existent source when `--merge` flag is used.
 
 ## [4.2.0] - 2025-04-04
 ### Changed